--- products/spiders/harrisfarm_au.py
@@ -1,23 +1,0 @@
-from scrapy.spiders import SitemapSpider
-
-from products.structured_data_spider import StructuredDataSpider
-
-class HarrisFarmAUSpider(SitemapSpider, StructuredDataSpider):
-    name = "harrisfarm_au"
-    allowed_domains = ["www.harrisfarm.com.au"]
-
-    sitemap_urls = ["https://www.harrisfarm.com.au/robots.txt"]
-    sitemap_rules = [
-        (r"https://www.harrisfarm.com.au/products/[\w-]+", "parse_sd"),
-    ]
-{'@context': 'http://schema.org/', 
-'@type': 'Product', 
-'name': 'Simply Stirred White Balsamic Blueberry Dressing 250ml', 
-'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594', 
-'image': ['https://www.harrisfarm.com.au/cdn/shop/files/96594.png?v=1725251792&width=1920'], 
-'description': 'Simply Stirred White Balsamic Blueberry Dressing\xa0\nSavour the soft, gentle flavour of white balsamic vinegar caramelised into a luxurious dressing.\xa0 The addition of blueberries gives this dressing a unique richness as well as its striking colour.\xa0\nGreat added to pan juices when cooking meats to add a beautiful, glazed coating to your finished meal.\xa0 Also, fantastic dressing your favourite salad or green vegetable mix!', 
-'sku': '96594', 
-'brand': {'@type': 'Brand', 'name': 'HFM'}, 
-'offers': [{'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/OutOfStock', 'price': 3.99, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822413226043'},
-
- {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/OutOfStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822413258811'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822413291579'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822413324347'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822413357115'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822413389883'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822413422651'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822413455419'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822413488187'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822413520955'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822413553723'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822413586491'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822413619259'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822413652027'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822413684795'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822413717563'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822413750331'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822413783099'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822413815867'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/OutOfStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=42033294213179'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822413848635'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822413881403'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822413914171'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822413946939'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822413979707'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822414012475'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822414045243'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822414078011'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822414110779'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822414143547'}, {'@type': 'Offer', 'sku': '96594', 'availability': 'http://schema.org/InStock', 'price': 11.49, 'priceCurrency': 'AUD', 'url': 'https://www.harrisfarm.com.au/products/simply-stirred-white-balsamic-blueberry-dressing-250ml-96594?variant=41822414176315'}]}
